# apps/api/src/assets/api.yaml
openapi: 3.0.3
info:
  title: Dataset Register
  description: Dutch Digital Heritage Network Dataset Register REST API
  version: 'latest'
  contact:
    name: Dutch Digital Heritage Network
    url: 'https://datasetregister.netwerkdigitaalerfgoed.nl/faq.php'
    email: bob.coret@netwerkdigitaalerfgoed.nl
servers:
  - url: 'https://datasetregister.netwerkdigitaalerfgoed.nl/api'
    description: 'Dataset Register API'
  - url: 'http://localhost:3000'
    description: 'Local development Dataset Register API'
paths:
  /datasets:
    post:
      summary: Add dataset description to the Dataset Register
      description: |-
        Submit dataset description(s) to the Dataset Register. Each dataset description will be validated before it is added to the Dataset Register.

        Only URLs that are on the allow list can be registered. Please [contact us](mailto:bob.coret@netwerkdigitaalerfgoed.nl) to have your domain added.
      requestBody:
        content:
          application/ld+json:
            schema:
              type: object
              required:
                - '@id'
              properties:
                '@id':
                  type: string
                  description: 'URL where the dataset description(s) can be found.'
            examples:
              valid:
                summary: A valid dataset
                value: { '@id': 'https://example.com/dataset' }
              invalid:
                summary: An invalid dataset
                value: { '@id': 'https://example.com/invalid-dataset' }
      responses:
        202:
          description: All dataset descriptions at the submitted URL are valid according to the <a href="https://netwerk-digitaal-erfgoed.github.io/requirements-datasets/">Requirements for Datasets</a>. The datasets will added to the Dataset Register shortly.
          content:
            application/ld+json: {}
        400:
          description: One or more dataset descriptions are invalid according to the <a href="https://netwerk-digitaal-erfgoed.github.io/requirements-datasets/">Requirements for Datasets</a>. The response body contains a list of [SHACL](https://www.w3.org/TR/shacl/) violations.
        403:
          description: The submitted URL is not on the domain name allow list. [Contact us](mailto:bob.coret@netwerkdigitaalerfgoed.nl) to have your institution’s domain name added.
        404:
          description: The URL cannot be resolved.
        406:
          description: The URL can be resolved but it contains no datasets.
    delete:
      summary: Delete a dataset registration
      description: |-
        Delete a registration from the Dataset Register, including all dataset descriptions
        that were fetched from the registered URL. This operation requires authentication
        via a Bearer token.
      security:
        - bearerAuth: []
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: Registration URL to delete
          example: 'https://example.com/catalog'
      responses:
        204:
          description: Registration and all linked datasets have been deleted successfully.
        400:
          description: Missing or invalid URL parameter.
        401:
          description: Missing or invalid Authorization header.
        404:
          description: No registration found for the given URL.
  /datasets/validate:
    post:
      summary: Validate dataset description(s) in the request body
      description: |-
        Validate dataset description(s) according to the <a href="https://netwerk-digitaal-erfgoed.github.io/requirements-datasets/">Requirements for Datasets</a>.
      operationId: validate-body
      requestBody:
        content:
          text/turtle:
            schema:
              type: string
            examples:
              valid:
                summary: A valid dataset in the request body
                value: |-
                  @prefix schema: <https://schema.org/> .
                  <http://data.bibliotheken.nl/id/dataset/rise-alba>
                    a schema:Dataset ;
                    schema:name "Alba amicorum van de Koninklijke Bibliotheek" ;
                    schema:license <http://creativecommons.org/publicdomain/zero/1.0/> ;
                    schema:publisher <https://example.com/publisher> .

                  <https://example.com/publisher>
                    a schema:Organization ;
                    schema:name "Koninklijke Bibliotheek" .
          application/ld+json:
            schema:
              type: object
            examples:
              valid:
                summary: A valid dataset in the request body
                value:
                  {
                    '@context': 'https://schema.org/',
                    '@type': 'Dataset',
                    '@id': 'http://data.bibliotheken.nl/id/dataset/rise-alba',
                    'name': 'Alba amicorum van de Koninklijke Bibliotheek',
                    'license': 'http://creativecommons.org/publicdomain/zero/1.0/',
                    'publisher':
                      {
                        '@type': 'Organization',
                        '@id': 'https://example.com/publisher',
                        'name': 'Koninklijke Bibliotheek',
                      },
                  }
      responses:
        200:
          $ref: '#/components/responses/Valid'
        400:
          $ref: '#/components/responses/Invalid'

    put:
      summary: Validate dataset description(s)s at a URL
      description: |-
        Validate dataset description(s) according to the <a href="https://netwerk-digitaal-erfgoed.github.io/requirements-datasets/">Requirements for Datasets</a> that are available at the submitted URL. That URL may resolve:

        - a single dataset which will be validated;
        - multiple datasets, each of which will be validated;
        - multiple datasets in a data catalog, which will be validated in full.
      operationId: validate
      requestBody:
        content:
          application/ld+json:
            schema:
              type: object
              required:
                - '@id'
              properties:
                '@id':
                  type: string
                  description: 'URL where the dataset description(s) can be found.'
            examples:
              valid:
                summary: A valid dataset
                value: { '@id': 'https://example.com/dataset' }
              invalid:
                summary: An invalid dataset
                value: { '@id': 'https://example.com/invalid-dataset' }
      responses:
        200:
          $ref: '#/components/responses/Valid'
        400:
          $ref: '#/components/responses/Invalid'
        404:
          description: The URL cannot be resolved.
        406:
          description: The URL can be resolved but it contains no datasets.
  /allowed-domains:
    get:
      summary: Check whether a URL’s domain is on the allow list
      description: |-
        Check whether the domain of the given URL is on the allow list.
        Only URLs on allowed domains can be registered via POST /datasets.
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: URL whose domain to check
          example: 'https://example.com/dataset.ttl'
      responses:
        200:
          description: The URL’s domain is on the allow list.
        400:
          description: Missing or invalid URL parameter.
        404:
          description: The URL’s domain is not on the allow list.
    post:
      summary: Add a domain to the allow list
      description: |-
        Add a domain to the allow list, so that URLs on that domain (or any of its subdomains)
        can be registered via POST /datasets. This operation requires authentication
        via a Bearer token.

        The domain must be a registrable domain (e.g. `example.com`), not a subdomain
        or full URL. Adding a domain that is already on the allow list is a no-op.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - domain
              properties:
                domain:
                  type: string
                  description: Registrable domain to add to the allow list.
            examples:
              valid:
                summary: Add a domain
                value: { 'domain': 'example.com' }
      responses:
        204:
          description: The domain has been added to the allow list.
        400:
          description: Missing or invalid domain.
        401:
          description: Missing or invalid Authorization header.
  /shacl:
    get:
      summary: Get the SHACL shape that is used to validate dataset descriptions.
      description: |-
        Get the <a href="https://www.w3.org/TR/shacl/">SHACL shape graph</a> that is used to validate dataset descriptions.
      responses:
        200:
          description: A SHACL shape
          content:
            application/ld+json:
              example: |-
                [
                  {
                    "@id": "http://terms.netwerkdigitaalerfgoed.nl/ns/register#DatasetShape",
                    "@type": [
                      "http://www.w3.org/ns/shacl#NodeShape"
                    ],
                    "http://www.w3.org/ns/shacl#name": [
                      {
                        "@value": "Datasetbeschrijving",
                        "@language": "nl"
                      },
                      {
                        "@value": "Dataset description",
                        "@language": "en"
                      }
                    ]
                  }
                ]
            text/turtle:
              example: |-
                <http://terms.netwerkdigitaalerfgoed.nl/ns/register#DatasetShape> a <http://www.w3.org/ns/shacl#NodeShape>.
                    <http://www.w3.org/ns/shacl#name> "Datasetbeschrijving"@nl, "Dataset description"@en;
                    <http://www.w3.org/2000/01/rdf-schema#comment> "Een datasetbeschrijving bestaat uit een elementen die de dataset beschrijft"@nl;
                    <http://www.w3.org/ns/shacl#targetClass> <http://schema.org/Dataset>;
                    <http://www.w3.org/ns/shacl#nodeKind> <http://www.w3.org/ns/shacl#IRI>;
                    <http://www.w3.org/ns/shacl#property> <http://terms.netwerkdigitaalerfgoed.nl/ns/register#SchemaNameProperty>, <http://terms.netwerkdigitaalerfgoed.nl/ns/register#SchemaLicenseProperty>, <http://terms.netwerkdigitaalerfgoed.nl/ns/register#SchemaDescriptionProperty>, <http://terms.netwerkdigitaalerfgoed.nl/ns/register#SchemaPublisherPrope

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API access token for administrative operations
  responses:
    Valid:
      description: |-
        All dataset descriptions are valid according to the <a href="https://netwerk-digitaal-erfgoed.github.io/requirements-datasets/">Requirements for Datasets</a>.
        Note that `sh:conforms` is different from the dataset description’s validity: `sh:conforms` is `true` only if no violations, warnings or infos are returned.
        A dataset description is considered valid if it has no violations, even if warnings are present. 
        However, warnings cause `sh:conforms` to be `false`.
      content:
        application/ld+json:
          example:
            - '@id': '_:report'
              '@type':
                - 'http://www.w3.org/ns/shacl#ValidationReport'
              'http://www.w3.org/ns/shacl#conforms':
                - '@value': 'true'
                  '@type': 'http://www.w3.org/2001/XMLSchema#boolean'
        text/turtle:
          example: |-
            _:report a <http://www.w3.org/ns/shacl#ValidationReport>;
              <http://www.w3.org/ns/shacl#conforms> true.
    Invalid:
      description: One or more dataset descriptions are invalid according to the <a href="https://netwerk-digitaal-erfgoed.github.io/requirements-datasets/">Requirements for Datasets</a>. The response body contains a list of [SHACL](https://www.w3.org/TR/shacl/) violations.
      content:
        application/ld+json:
          example:
            - '@id': '_:b7'
              '@type':
                - 'http://www.w3.org/ns/shacl#ValidationResult'
              'http://www.w3.org/ns/shacl#resultSeverity':
                - '@id': 'http://www.w3.org/ns/shacl#Violation'
              'http://www.w3.org/ns/shacl#sourceConstraintComponent':
                - '@id': 'http://www.w3.org/ns/shacl#MinCountConstraintComponent'
              'http://www.w3.org/ns/shacl#sourceShape':
                - '@id': '_:df_28_15'
            - '@id': '_:df_28_15'
              'http://www.w3.org/ns/shacl#path':
                - '@id': 'http://schema.org/publisher'
              'http://www.w3.org/ns/shacl#minCount':
                - '@value': '1'
                  '@type': 'http://www.w3.org/2001/XMLSchema#integer'
              'http://www.w3.org/ns/shacl#class':
                - '@id': 'http://schema.org/Organization'
              'http://www.w3.org/ns/shacl#node':
                - '@value': 'schema:PublisherShape'
            - '@id': '_:b7'
              'http://www.w3.org/ns/shacl#focusNode':
                - '@id': 'http://data.bibliotheken.nl/id/dataset/rise-alba'
              'http://www.w3.org/ns/shacl#resultPath':
                - '@id': 'http://schema.org/publisher'
              'http://www.w3.org/ns/shacl#resultMessage':
                - '@value': Less than 1 values
            - '@id': '_:report'
              '@type':
                - 'http://www.w3.org/ns/shacl#ValidationReport'
              'http://www.w3.org/ns/shacl#conforms':
                - '@value': 'false'
                  '@type': 'http://www.w3.org/2001/XMLSchema#boolean'
              'http://www.w3.org/ns/shacl#result':
                - '@id': '_:b10'
        text/turtle:
          example: |-
            _:b11 a <http://www.w3.org/ns/shacl#ValidationResult>;
              <http://www.w3.org/ns/shacl#resultSeverity> <http://www.w3.org/ns/shacl#Violation>;
              <http://www.w3.org/ns/shacl#sourceConstraintComponent> <http://www.w3.org/ns/shacl#MinCountConstraintComponent>;
              <http://www.w3.org/ns/shacl#sourceShape> _:df_28_15.
            _:df_28_15 <http://www.w3.org/ns/shacl#path> <http://schema.org/publisher>;
              <http://www.w3.org/ns/shacl#minCount> 1;
              <http://www.w3.org/ns/shacl#class> <http://schema.org/Organization>;
              <http://www.w3.org/ns/shacl#node> "schema:PublisherShape".
            _:b11 <http://www.w3.org/ns/shacl#focusNode> <http://data.bibliotheken.nl/id/dataset/rise-alba>;
              <http://www.w3.org/ns/shacl#resultPath> <http://schema.org/publisher>;
              <http://www.w3.org/ns/shacl#resultMessage> "Less than 1 values".
            _:report a <http://www.w3.org/ns/shacl#ValidationReport>;
              <http://www.w3.org/ns/shacl#conforms> false;
              <http://www.w3.org/ns/shacl#result> _:b11.